Troubleshooting Jung 3D
This doesn't work. Yet. As I've been poking at it, it seems like it's got something to do with Java3D (which is library JUNG uses to render the 3D visualization). Jung includes the core of Java3D, but not the utilities, and it includes an older version. Compiling results in an error in which some of the classes defined in the utilities are not defined (duh: the utilities aren't there).

If you try to swap in a newer version's utilities, you run into a "sealing" error — that is, the libraries are sealed to a particular version and mixing and matching is blocked.

Completely replacing the older Java3D libraries included with JUNG still results in a sealing error, which suggests that Jung itself may depend on the older version. However, there is no native Mac JAR of the version JUNG uses, so we may need to do our own build, which may (or may not) get a trifle involved.

Signed-off-by: Seth Battis <SethBattis@stmarksschool.org>