Run a Command
Simplified the opening of an image to just use the ImageJ app. This may introduce some limitations — such as needing to have the image window open to run processing commands on it. But it streamlines the code, and we probably want windows open so that we can debug, for now.

Also demonstrated calling a command on the opened image in ImageJ, converting it to 8-bit color.

Albert Cardona has a [a good set of tutorials](http://albert.rierol.net/imagej_programming_tutorials.html) that got me here. Not directly, but they gave me enough hints to guess this.

Signed-off-by: Seth Battis <SethBattis@stmarksschool.org>