Mytstified
Apparently what works is to use the current version of JOGL, [this version of Java3D](http://gouessej.wordpress.com/2012/08/01/java-3d-est-de-retour-java-3d-is-back/) and to not include jung-3d.jar in the JUNG user library, but rather to add it as an external JAR... which causes it to _also_ include j3d-core-1.3.1.jar and vecmath-1.3.1.jar from that same build path.

But it works.

Signed-off-by: Seth Battis <SethBattis@stmarksschool.org>