Reordered Build Order
I think this is a deck chairs on the Titanic issue (that is: it's more important to get rid of old version of Java3D from within the buildpath than it is to get this order right). But the buildpath really should be identical to the dependencies in the program.

Signed-off-by: Seth Battis <SethBattis@stmarksschool.org>